Discover peers dynamically with Kademlia
Resolves #4, #5 and #15.

Note that we also keep mDNS as it provides a simple way of
starting a local-only network without needing to know the
addresses of bootstrap peers.

This also has a couple of related changes.

Firstly, we've re-structured the in-memory representation of the
committee. The three `Vec`s all had the same length, so we've
combined them into a single `Vec` of tuples to make that explicit.

Secondly, we resolve the node identity issue by storing a node's
BLS public key in the Kademlia DHT. Other nodes query the public
key from the DHT when they first discover a new node.