Configure nodes with a TOML configuration file (#48)
We choose to still pass the node's secret key on the command line,
to discourage operators from storing their secret key in plaintext
on disk.

Also note that all configuration parameters currently have
defaults, so the empty configuration file is a valid one. Therefore,
we add a special case which assumes that if the config file doesn't
exist, we assume the empty config file. This property may not live
forever and so we might undo this special case later.