Export `rpc.server.duration` metric via OpenTelemetry
The metric and its attributes aim to follow the OpenTelemetry
semantic versions, as documented at:
https://github.com/open-telemetry/semantic-conventions/blob/main/specification/metrics/semantic_conventions/rpc-metrics.md#rpc-server

This also adds the infrastructure for instrumenting other bits of
the code with OTel metrics.

We also add an OTel collector, Mimir and Grafana to
`docker-compose.yaml`, so developers should get a mostly working
instrumentation stack by default.