Move signature generation and verification into messages
Instead of generating and verifying signatures externally, we add
inherent methods on each message and perform the operations there
instead.

This centralises the logic for how signatures are generated.

In future, we may also wish to take this a step further and
disallow access to the messages' fields until the signature has
been verified.