Fix state root calculation
Previously, the transactions in block `n` would only be applied
to the state when the block was voted on, meaning block `n+1`
contained the state root with these transaction executions
included.

Now, transactions in block `n` are applied to the state by the
proposer immediately and the root hash is included in the block
before it is proposed to the network.

This also makes it trivial to start verifying state root hashes
when we vote on blocks, so this commit also resolves #84.