Fix double-execution of transactions by block proposer
Previously, we would execute transactions against our state both
when forming and proposing a block and also when voting on the
same block.

Instead, we remove the transaction from `new_transactions` whenever
it is executed, guaranteeing we will only execute it once.

There are probably problems here if a node proposes a block, but
the network does a view change and does not accept the block. In
this case, the node may need to rollback its state changes, but I
am deferring these problems for later.