Simplify transaction execution
We remove the `pending_tranasctions` list from `Consensus`, because
it was basically redundant with the list of keys in
`new_transactions` anyway. When constructing a block to propose, we
just use everything in `new_transactions`. This contains all
transactions we've been broadcasted, but haven't yet executed.

This also nicely resolves #82 since all nodes should have the same
list of `new_transactions` (ignoring latency of transaction
broadcasts).