Fix block hash stored in transactions
In #143, we made sure transactions were included in the block that
said they included them. However, we did not change the block hash
that is stored in each transaction. This meant that the block hash
was invalid and pointed to the parent block.

In order to fix this, we restructure things slightly to defer the
construction of `Transaction` until all the `NewTransaction`s in a
block have been executed. At this point we know the block hash and
so can construct the `Transaction`.