Fix double-execution of transactions and simplify execution logic (#150)
* Fix double-execution of transactions by block proposer

Previously, we would execute transactions against our state both
when forming and proposing a block and also when voting on the
same block.

Instead, we remove the transaction from `new_transactions` whenever
it is executed, guaranteeing we will only execute it once.

There are probably problems here if a node proposes a block, but
the network does a view change and does not accept the block. In
this case, the node may need to rollback its state changes, but I
am deferring these problems for later.

* Simplify transaction execution

We remove the `pending_tranasctions` list from `Consensus`, because
it was basically redundant with the list of keys in
`new_transactions` anyway. When constructing a block to propose, we
just use everything in `new_transactions`. This contains all
transactions we've been broadcasted, but haven't yet executed.

This also nicely resolves #82 since all nodes should have the same
list of `new_transactions` (ignoring latency of transaction
broadcasts).