Implement EVM opcodes which query the current chain state
In particular, the EVM can now query the current block's timestamp
and number. I've left placeholder responses for a few other
operations.

To support this, I've added a timestamp to blocks. This gets
validated by voters based on the following 2 rules:
1. It must not have decreased from the parent block
2. It must be sufficiently close to the current time

It was also helpful to split `Block` into a `Copy`-able subset
that we can pass to the execution engine, without keeping a
borrow - This subset is the `BlockHeader`.