Store full transaction objects in blocks, rather than just transaction hashes.
While it makes the Block object larger, it should make network propagation more consistent as all transactions need to be gossiped regardless for the new block to be voted on. Including them directly greatly simplifies the logic rather than having to wait for every individual network message to arrive.

Additionally this adds an index of transactions to block hashes in consensus. The exact schema of our data will probably need to be looked at in more detail when adding persistent storage.