Normalise transaction storage in `Consensus`
We remove the duplication of committed `Transaction`s in `blocks`
and in `transactions`, by removing the full `Transaction` from
`Block` and instead just storing the hash.

We still want to keep the work of #157 which ensures we distribute
full transactions with each block proposal, so we slightly alter
the `Proposal` message to effectively be a `Block` plus a list of
`Transactions`.