Make block production test deterministic (#168)
We do this by removing the networking/libp2p layers and instead
exercising `Node` directly. The test function constructs a list of
nodes and manages a message stream from each of them. It probably
makes sense to factor this into a test harness at some point soon,
if we want to add more of these kind of tests.

Resolves #159.