Handle logs from EVM executions
When a contract execution produces logs, we retain them and return
them in the `eth_` APIs.

I have introduced the concept of `TransactionReceipt`s for storing
data about `Transaction`s which is not known until after they have
been executed. I moved the `block_hash` and `contract_address` into
the receipt, as well as the list of `Log`s.

I also did a bit of cleanup to remove specific types that are
specific to the Ethereum API from `Node`, so this commit also
resolves #78.