Refactor RPC module declarations
We move the `method!` macro from `eth.rs` into `api/mod.rs` so it
can be used by other APIs as well. Also, instead of exposing a
`method!` macro which expects various variables to be in scope and
is called multiple times, we instead expose a `declare_module!`
macro which should be called once with a list of methods and
returns a full `RpcModule`.

This also moves the single Zilliqa API method that we've
implemented to the same macro.

Finally, the `net_*` and `web3_*` methods are moved out of
`eth.rs` and into their own namespace-specific module.