Implement native token balances (#205)
As discussed in #81, we have chosen to implement the native token
with an ERC-20 contract that is implicitly called when the native
balance is updated.

I have chosen to include both the compiled EVM bytecode and the
Solidity source code in-tree. My justification for the former is
that it simplifies builds by avoiding needing to compile Solidity
at build-time and makes it easier to determine that builds are
deterministic. The Solidity source code is included, along with a
test case that downloads and verifies a specific version of the
Solidity compiler, runs it against the checked-in source and
validates the output corresponds to the checked-in bytecode. This
test is only run when a specific feature is enabled, but all
features are enabled when running tests in CI. Also note that the
native token contract relies on imports from the
openzeppelin-contracts repository, which I've added as a submodule.

The current solution also has a number of caveats and controversial
decisions:

* The EVM engine returns balance updates to us in the form "Update
account X to balance A", "Update account Y to balance B". This is
difficult to reconcile with the standard ERC-20 interface, which
only allows transfers between accounts. Therefore, I have extended
the contract to add a "setBalance(account, amount)" function.
* When a native balance is updated, we apply the changes to the
account state using a transaction, but the result (mainly the logs)
of that transaction is not persisted anywhere. This means the
balance of the ERC-20 contract gets updated, without any `Transfer`
logs being visible on the chain.
* There are a couple of 'hacky' shortcuts in the transaction
application code that are needed to avoid infinite recursion when
obtaining or updating a native balance:
  * If a tranasction is sent to the native token contract, it
can't update the native balance.
  * The native balance of the zero address and the native token
contract are returned to the EVM execution engine as zero.

Resolves #70 and resolves #81.