Add test case for `eth_sendRawTransaction` (#214)
* Add test case for `eth_sendRawTransaction`

This required some extra work on the test harness to obtain the
exposed `RpcModule` from nodes and to hook this up to ethers-rs.
The result is that we can write tests which use the convinient
ethers API, but no network requests are actually made and
everything stays within a single process.

Since we're exercising the transaction creation code via the API,
I decided to remove the existing transaction creation test. This
also meant that nothing was using `ManualConsensus` anymore, so
that is also removed.

Also, I've moved all the integration tests into a single `it`
(integration tests) module under the `tests` directory. This means
`rustc` only compiles a single binary for all our tests, reducing
the time on linking. As our test suite grows, the effects of this
should become more substantial:
https://matklad.github.io/2021/02/27/delete-cargo-integration-tests.html#Implications