Implement `ots_searchTransactions{Before,After}` (#210)
These are Otterscan API methods to search through the list of
transactions that 'touched' an account, where touched can mean:
* A transaction sent from the account
* A transaction sent to the account
* A transaction that sent money to or called the account as part
of its trace

To implement this we add an additional index to our nodes which
stores a map of addresses to transactions which touched the
address.

The rest of the code iterates over the index and populates the
response.

I considered an alternative index that is worth mentioning and we
might want to revisit in the future:

* An index of address to blocks in which this address was touched.
This is the approach taken by Erigon. It makes it much easier to
discount transactions which are outside of the range specified in
the API request, but requires you to execute every transaction in
the matched blocks to filter out the ones which don't touch the
address. I didn't opt for this, because we don't have the ability
to replay transactions yet (#158). Also, its easier to optimise
a sparse `Vec<u64>` than a `Vec<Hash>` with (for example) a
run-length encoded bitmap.