Implement `eth_getStorageAt` (#228)
* Implement `eth_getStorageAt`

In order to test this with the provided example, I added a helper
macro to the test suite to deploy contracts, given their source. I
decided it would be simplest to just compile them at runtime and I
figured we could revisit this decision quite easily later if that
turns out to be slow.