Refactor transaction types (#237)
We make the existing `Transaction` unsigned and add a new wrapper
type - `SignedTransaction` which includes a `Transaction` and a
`SigningInfo`. This allows us to represent both signed and
unsigned transactions at any point in our system.

There is some duplication of data in `SignedTransaction` - The
`from_addr` is a pure function of the `signing_info`, but we
choose to store it anyway to avoid having to repeat the expensive
address recovery process.

We also change the way we generate transaction signatures to match
Ethereum's and update a test case to check this.