Make transactions' `to_addr` optional
Instead of representing contract creations with the zero address,
we instead make the to address optional and represent creations
with a to address of `None`.

This is in fact how Ethereum works if you dig deep enough into the
yellow paper, though all the non-normative documentation is quite
confused on the subject.

Specifically, equation (20) in section 4.2 of the
[yellow paper](https://ethereum.github.io/yellowpaper/paper.pdf)
states that `T_t` (the to address of a transaction) "is either a
20-byte address hash or, in the case of being a contract-creation
transaction (and thus formally equal to the empty set), it is the
RLP empty byte sequence and thus the member of `B_0`.

This is important in practice, because a transaction with a to
address of `Some(0x0)` has a different hash (and thus signature)
than one with a to address of `None`. Previously, this would
cause our public key recovery to fail with one of these two
encodings, because we unified them.