Allow nodes to join the committee after the network has started
We add a new message type `JoinCommittee` which is broadcasted
by a new node when it wishes to join the consensus committee. When
a node receives a `JoinCommittee` message, it adds the specified
node to the committee at the next block it proposes.

To ensure the network has a consistent view of the committee at
any stage, we record the current committee inside each block.

To start the network from genesis, the genesis block requires a
committee with at least one member, who will be able to propose
and vote on new blocks, which can add further members to the
committee. Therefore, we add a `genesis_committee` configuration
value to set the committee of the genesis block.

When a node joins the network, it needs to download and verify the
existing chain of blocks from genesis. This is done by the
`BlockRequest` and `BlockResponse` messages, coordinated by the
new `BlockStore` structure.