Avoid double-wrapping jsonrpsee errors (#345)
Previously, we would wrap all errors in an `ErrorObject` with an
error code of `InternalError`. Sometimes our handlers return
`ErrorObject`s themselves already, such as when a parameter is
invalid. This meant all errors would be returned as an internal
error, even when we could be more specified.

Instead, we attempt to downcast the returned `anyhow::Error` and
avoid wrapping if it is already an `ErrorObject`.

Resolves #343.