Concurrent networks with contracts (#405)
You can now send a transaction to the shard registry to have the node automatically spawn a child shard.

Every chain has a shard contract at a well-defined address, containing consensus-relevant settings, and a function for validators to join the network. The main shard additionally contains a listing of all its child shards at the same address.
In the main chain node, every time we finalize a block, we scan its logs, and if we determine that a new shard was added we launch a sub-node for that shard.
This will likely be subject to change, but currently the design is relatively tentative. The main functionality added here is the framework for analysing block logs and spawning shards based on that.

One notable change done to this effect was to reorganise transaction receipt storage to be per-block, making it more straightforward to fetch every receipt (and hence log) in a given block.