Fix balance scaling for Zilliqa `GetBalance` API (#457)
Internally we represent balances with 18 decimal places. However,
the Zilliqa API expects 12 decimal places, so we need to scale and
truncate the amount.

I figured it made more sense to truncate rather than round,
because a lower bound for an account's balance is more useful if
you're trying to work out how much they could spend. Happy to
discuss if others disagree though.