Improve how contracts are pre-compiled and tested
Instead of a separate `.json` file per contract, we instead have a
single `compiled.json` which contains the output artifacts of all
contracts.

We've stopped using the deprecated `--combined-json` option of
`solc` and are instead using the `--standard-json` interface, which
is what gives us the combined JSON output. External tooling is
likely to have better support for this format.

We've also changed how the compilation assertion tests work.
Instead of requiring the developer to manually run `solc` and
check in the results, you are now expected to run the assertion
test with the `ZQ_CONTRACT_TEST_BLESS` environment variable set.
In this case, the test will recompile the contracts and update the
`compiled.json` file on disk. If the variable is not set, the test
compares the compiled value with the file on disk.

The final change is that we are using the `solc` support in
`ethers-rs` to compile in the test. Under the hood, this still
shells out to `solc`, but it uses `svm-rs` to download and manage
the correct version. This means the test should now work on
Windows and Mac, as well as Linux.