Abstract database serialization and access into separate module (#374)
All references to `sled` and (de)serialization code now live
inside `db::Db`, which exposes a simpler interface to `Consensus`,
`BlockStore` and `State`. These structs now store an `Arc<Db>`.

I renamed the existing `SledDb` to `TrieStorage` to avoid
confusion with my new struct and to hopefully make the purpose
clearer.

We make use of `paste` in the implementation of `Db` to cut down
on some repetition. Feel free to complain if you think this
doesn't help.