Support type 1 and type 2 Ethereum transactions (#512)
* Support type 1 and type 2 Ethereum transactions

I did some fairly major refactoring to make this support easier.
We now have three different transaction types, where we used to
have two:

* `Transaction` - The core information of a transaction. This is
now an enum over the possible transaction types, rather than a
struct which tries to unify the types. I've added accessor methods
to the enum to obtain the common properties of all variants.
* `SignedTransaction` - A `Transaction` plus a signature. The
details of the signature depend on the tranasction type. This is
also an enum over the possible transaction types.
* `RecoveredTransaction` - A `SignedTransaction` plus the
transaction's `signer` and its `hash`. This can be obtained by
calling `SignedTransaction::recover_signer()`. We deliberately
don't implement `serde` traits for this, to avoid it being sent
over the network. This ensures nodes need to validate transaction
signatures independently.

* Response to PR feedback

* Rename transaction types