Add flag to run test suite in a 'summary' mode (#594)
When running the test suite with `ZQ_TEST_SEED_SUMMARY` set, the
test suite will not fail fast and keep running if a seed fails.
The failing seeds will not be printed - Instead, a summary with
the total number of passes and failures will be printed.

My intention was that this could be run with a high seed count,
outside of the normal PR process. This would let us tell if our
suite is flaky and, if so, how flaky.