Pr feedback on scilla integration (#690)
* Support type 1 and type 2 Ethereum transactions

I did some fairly major refactoring to make this support easier.
We now have three different transaction types, where we used to
have two:

* `Transaction` - The core information of a transaction. This is
now an enum over the possible transaction types, rather than a
struct which tries to unify the types. I've added accessor methods
to the enum to obtain the common properties of all variants.
* `SignedTransaction` - A `Transaction` plus a signature. The
details of the signature depend on the tranasction type. This is
also an enum over the possible transaction types.
* `RecoveredTransaction` - A `SignedTransaction` plus the
transaction's `signer` and its `hash`. This can be obtained by
calling `SignedTransaction::recover_signer()`. We deliberately
don't implement `serde` traits for this, to avoid it being sent
over the network. This ensures nodes need to validate transaction
signatures independently.

* Response to PR feedback

* WIP

* pause before merging james' tx changes

* wip

* wip

* wip

* WIP - compiles with rt block

* wip - need to connect through tcp serv

* more progress

* WIP, more progress. scilla unhappy with response

* refactored

* refactoring

* wip

* wip

* wip

* cleanup code, delete commented out code

* more cleanup

* merge to master, update readme

* cleaned up tcp read to dynamically allocate buffer

* save before updating the call scilla response type

* successfully passes the tests for deploy and call hello world contract

* updated for b64 support

* more cleanup

* Update README.md

* re-enable tcp

* more cleanup

* more cleanup

* cleanup

* WIP

* cleanup

* wip

* clean up to get that green tick

* fix h2

* dependency updates

* linting

* WIP

* WIP

* major rafactoring, works fine

* another refactor, works

* more refactoring

* put a proper loop into the scilla handler (but no transfers supported atm)

* more refactoring

* save before cleaning up for master merge

* minor update, getting ready for master merge

* linting

* linting

* linting

* more linting

* remove submodule

* refactor

* linting

* debug odd reason build failed just now

---------

Co-authored-by: James Hinshelwood <jameshinshelwood1@gmail.com>