Fix block number clamping when requesting new blocks (#717)
The current version of the code clamps the requested block to the
current head block's number. This means when we realise we've
missed blocks, a first request is made to download blocks between
the current head and the latest known one. Once the response is
recieved and the blocks are processed, we make a second request
for the latest block again. The response contains this block and
we ignore it, because we already have it.

With the new version of the code, we request the latest block plus
one. This results in an empty block response, assuming the next
block doesn't yet exist.

This is still necessary, because in the current version of the
block request/response protocol, we require an empty (or rejected)
response to indicate there are no more blocks to download. In
future, we should change the protocol to indicate this more
efficiently.