Change EVM implementation from SputnikVM to REVM (#750)
REVM is more actively supported and maintained and the codebase
is simpler and easier to understand, in my opinion.

A few caveats and notes about things I've temporarily broken to
make the size of this change more manageable:
* Scilla support is removed for now - The communication protocol
was suboptimal and needs to be rewritten anyway.
* The `ots_getInternalOperations` API was removed - It didn't work
properly so we need to work on it soon in any case.
* Account balances are no longer represented in an ERC-20 - Instead
we've agreed to implement a precompile that exposes the native
balances as an ERC-20.