Fix response format of `GetVersion` (#790)
It now returns an object containing `Commit` and `Version` fields,
which is what ZQ1 does.

`Commit` is populated by the full SHA of the git commit at which
the node was built.

`Version` is populated by a git tag, if one exists or an
abbreviated commit SHA if not.
If the working directory is dirty, `-dirty` is appended to the
version.
This is close to the output of
`git describe --always --dirty --tags`.

Example output when `HEAD` is tagged:
```
{
    "Commit": "c2ecdc27e588a0ababcbe9a00567351d739664b4",
    "Version": "foobar"
}
```

Example output when `HEAD` is not tagged:
```
{
    "Commit": "c2ecdc27e588a0ababcbe9a00567351d739664b4",
    "Version": "c2ecdc2"
}
```