Implement `eth_subscribe` for `newHeads` (#819)
This adds the `eth_subscribe` and `eth_unsubscribe` APIs for the
Ethereum pub-sub protocol.

Currently, only `newHeads` are supported (`logs` and
`newPendingTransactions` are also specified). The API only works
over a websocket, because they need a duplex connection.

The implementation works by adding a `broadcast::Sender` to
`Consensus` which publishes `BlockHeader`s whenvever they are
produced. When a client makes a new subscription, it subscribes to
this sender. The sender keeps a buffer of the 4 latest blocks in
memory. If any reciever falls behind these 4 blocks (because
sending the blocks to the client is too slow), it will be
disconnected.