Implement Scilla contract creation and calls (#791)
* Implement Scilla contract creation and calls

`Account` used to have `code: Vec<u8>` and `storage_root: Option<H256>`.
Now it has a `contract: Contract`, where `Contract` is an `enum` with `Evm` and `Scilla` variants.
The `Evm` variant has the same fields as the old accounts.
The `Scilla` variant has `code: String`, `init_data: String` and `storage: HashMap<String, (ScillaValue, String)>` for representing Scilla contracts.
EOAs are represented as `Evm` contracts with empty code and a `None` storage root (as they used to be).

Transaction logs are now an enum with `Evm` and `Scilla` halves.
Since interop is not yet implemented, `Evm` logs are only produced by EVM transactions and `Scilla` logs are only produced by Zilliqa transactions, however this will eventually change.
The API layer filters out logs that are not of the expected type (i.e. only `Evm` logs are included in the response of `eth_getLogs`).
This will also change in the future - We've already agreed on an encoding for EVM logs in the Scilla API and vice versa.

The bulk of the logic is in `scilla.rs`, which implements the Scilla interface.
The comments in that file are the best way to understand what is going on.
One point worth noting here is that we introduce a runtime dependency for the test suite on `docker` - Each group of tests will start a container which runs the OCaml Scilla server.

* Respond to PR feedback

* Avoid using files to communicate with Scilla

Thanks to the latest Scilla update, we now send all contracts and
messages in the HTTP request body, rather than writing them to
disk.