Fix precision of account balances provided to Scilla (#848)
Previously, when the Scilla interpreter queried an account's
balance, we would return the result in units of (10^-18) ZILs.
Scilla expects the result in units of (10^-12) ZILs.

To mitigate similar bugs in the future, I have introduced a
newtype for quantities of ZIL expressed in units of (10^-12) ZILs:
`ZilAmount`. The intention is that we continue to use `u128` in
most places (e.g. for internal storage of amounts and for EVM
transactions), where we continue to expressed amounts as (10^-18)
ZILs. But for places we know we want amounts expressed as (10^-12)
ZILs (such as the Zilliqa API requests and responses), we use
`ZilAmount` to make that clear. A `ZilAmount` can't be converted
to a `u128` without multiplying it by `(10^6)` and vice-versa.