Add 'deployer' tool to perform network upgrades (#862)
Currently we have two commands:

* `new <name> <gcp_project> <gcs_bucket>` creates the TOML
configuration file for a network. This is used by other commands
and intended to be checked into git somewhere.
* `upgrade <config_file>` upgrades a network, based on the values
from the configuration file. It will build a ZQ2 binary at the
configured revision, upload it to GCS and upgrade the binaries on
each VM instance.

Co-authored-by: James Hinshelwood <jameshinshelwood1@gmail.com>