Refactor Scilla transaction execution (#916)
Previously when we applied a Scilla transaction to state (via
`State::apply_transaction_scilla`), the state would be mutated
directly by execution engine while the transaction was run. This
differed from the behaviour of EVM transactions, where the
execution engine only read from the state and kept changes to
state in memory (in a `HashMap<Address, Account>`). This PR
updates Scilla execution to work in the same way.

The advantage of this approach is that we get a state diff when
executing a transaction, which makes implementing APIs similar to
`trace_replayTransaction` possible. It also avoids more roundtrips
from/to the database, since each account is loaded at most once
and then saved once at the end of execution.