Improve block download speed when syncing with a network
There are quite a few changes here but the motivation for all of
them is to allow nodes to request multiple blocks in parallel.

The previous behaviour was that our `Consensus` was (seemingly)
randomly scattered with calls to `download_blocks_up_to_head`.
This would send out a `BlockRequest` for the current height plus
one. Nodes would respond to these requests with the requested
block, plus up to 100 additional blocks in the same batch.

All of this code has been removed and the logic is mostly
handled in `BlockStore`. We now inform `BlockStore` when we
see blocks that don't form a chain with our existing blocks.
It buffers these proposals in case they can form a chain later
and keeps track of the `highest_known_view` it has seen. When
it is poked to do so (after a new block proposal is received),
it sends out requests for as many blocks as possible,
respecting its configuration for the batch size and maximum
number of in-flight blocks. Importantly, these requests are
now sent and handled in parallel.

Other tangential changes made in support of this:
* Block requests and responses have been simplified to a
single type for each. The non-batch versions were unused
anyway.
* `genesis_hash` is removed from configuration since it
isn't used anywhere and adds complexity.
* We now log executed SQL statements.

Do feel free to complain if these changes make the PR
too big and you want me to split them up.