Fix transaction deserialization
Previously, our `serialize` and `deserialize` implementations were
incompatible. The `serialize` implementation uses
`serialize_bytes`. The `deserialize` implementaion delegated to
`Vec<u8>`'s implementaion, which calls `deserialize_seq`
internally. In JSON, these resulted in the same serialized form
but the same is not true for CBOR, which specialises
`serialize_bytes` to be more efficient.

Now we correctly call `deserialize_bytes` in the `deserialize`
implementation.