Fix transaction deserialization (#1077)
The changes in #1061 fixed (de)serialization for CBOR, but broke
it for JSON. This is because `serde_json` serialises bytes as an
array by default. At deserialisation, it does not know this array
represents bytes and so `visit_seq` is called. Previously, our
visitor returned an error here, because it was only expecting
`visit_bytes` to be called. Now we accept both.