Add 'blessed' transactions which are exempt from gas price rules (#1072)
* Currently this just includes the deployment transaction for the deterministic deployment proxy.

This PR also enables the REVM balance check. Previously, this was disabled because Scilla transfers would use the EVM execution flow and they often contained overly high gas limits. This is no longer the case. More alarmingly, it turns out that when the check is disabled, REVM just gives the sender of a transaction enough money to cover that transaction, meaning funds are invented out of thin air. So now mean that the sender of a transaction needs to have enough money to cover the provided `gas_limit * gas_price + amount`.

This means we've had to change how our 'system' transactions are processed. Previously, they would set a very high gas_limit and gas_price and the signer was the zero address. This obviously doesn't work any more since the zero address doesn't have any money at genesis. Instead, we now set the gas_price to zero and disable the gas price checks.

Resolves #1059

* Use enum for validation flag instead of bool