Specify external address in configuration
In the proto-testnet we are observing nodes learning about the
wrong external address for themselves, reporting the wrong thing
and thus other nodes failing to send messages to them.

I believe this is due to a bug in `libp2p-identify` where incoming
TCP ports are reported as the observed address of a node. However,
I think (and the documentation states) we shouldn't be relying on
other nodes to tell us our external address anyway since it is
easily exploitable by a malicious node.

So for now, we let configuration specify a nodes external address
and surpress the existing behaviour in this case.

In future, we can remove the configuration requirement by using
`libp2p-autonat` (#1101).